context/Theme/ThemeProvider.tsx
"use client";

import React, { createContext, useState, useEffect } from "react";
import { handleThemeChange } from "./utils";
import { ThemeContextType } from "./types";

export const ThemeContext = createContext<ThemeContextType | undefined>(
  undefined
);

export function ThemeProvider({ children }: { children: React.ReactNode }) {
  const [mode, setMode] = useState("");

  useEffect(() => {
    handleThemeChange(mode, setMode);
  }, [mode]);

  return (
    <ThemeContext.Provider value={{ mode, setMode }}>
      {children}
    </ThemeContext.Provider>
  );
}
